pyproject.toml
[project]
name = "rint-data-manager"
version = "0.1.0"
description = "DVC-based data collection and sharing platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "dvc>=3.0.0",
    "aiofiles>=23.2.1",
    "python-dotenv>=1.0.0",
    "email-validator>=2.3.0",
    "jinja2>=3.1.6",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0"
]
